APMXGeneRTCLib: Get base address from PcdApmRtcBase
From: Mark Salter <msalter@redhat.com>

The existing PcdApmRtcBase is not used by the RTC driver. Furthermore it is
not properly defined (address value in token field). This patch fixes the
definition of PcdApmRtcBase and has the RTC driver use it to get the base
address of the RTC registers.

Signed-off-by: Mark Salter <msalter@redhat.com>
Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>