APMXGeneRTCLib: fix RTC driver runtime support for OS
From: Mark Salter <msalter@redhat.com>

The existing XGene RTC driver causes a Linux kernel Oops:

  Unable to handle kernel paging request at virtual address 10510000
  pgd = ffffffc00007d000
  [10510000] *pgd=00000043e801c003, *pmd=0000000000000000
  Internal error: Oops: 96000005 [#1] SMP
  ...
  Call trace:
  [<ffffffc3fade3840>] 0xffffffc3fade3840
  [<ffffffc0003f50a8>] platform_drv_probe+0x24/0x58
  [<ffffffc0003f36ec>] driver_probe_device+0x90/0x370
  [<ffffffc0003f3abc>] __driver_attach+0x9c/0xa0
  [<ffffffc0003f18dc>] bus_for_each_dev+0x58/0x98
  [<ffffffc0003f316c>] driver_attach+0x20/0x28
  [<ffffffc0003f2d5c>] bus_add_driver+0x168/0x240
  [<ffffffc0003f3eb8>] driver_register+0x68/0x130
  [<ffffffc0003f507c>] __platform_driver_register+0x60/0x68
  [<ffffffc0003f5120>] platform_driver_probe+0x2c/0xac
  [<ffffffc000909634>] efi_rtc_driver_init+0x1c/0x24
  [<ffffffc0000813fc>] do_one_initcall+0xc4/0x14c
  [<ffffffc0008d69d4>] kernel_init_freeable+0x1cc/0x26c
  [<ffffffc00063b848>] kernel_init+0x10/0x104

The faulting address is the physical address of the RTC register base.
This Oops happens after SetVirtualAddressMap() has been called. The
UEFI RTC driver is not creating a runtime MMIO region in the memory
map and not remapping the register base address in response to the
SetVirtualMemoryMap() event. This patch fixes these problems so that
the kernel no longer Oopses.

Signed-off-by: Mark Salter <msalter@redhat.com>