Turn on Mustang UART FIFOs
The Mustang UART driver was not turning on UART FIFOs. This
led to broken VT100 screen control in grub. For example, the
arrow keys could not be used to select a boot entry in the
grub menu because the 3-character sequence generated by the
arrow keys would overrun the UART.

Signed-off-by: Mark Salter <msalter@redhat.com>