fix secondary core boots on mustang
For some reason, APM changed the secondary core pen address
(KERNEL_MAILBOX_ADDR) in their 1.12 UEFI release. This new
address no longer matches our device tree and thus causes
failure to boot the secondary cores in the kernel. We could
fix the device tree, but we have a mix of systems with various
versions of firmware and software. The safest thing to do is
to just change the address back to where it was. At some point,
ACPI will make this moot anyway.

Signed-off-by: Mark Salter <msalter@redhat.com>