[mustang] give sensible defaults to UpgradeFirmware.efi
If UpgradeFirmware is called without a file argument (or with extra
arguments), it will try to open a file called apm_upgrade.cmd. If that
works, it will execute the instructions from the file. Otherwise, it
will print its usage instructions and fail.

Also, fix a minor typo: "writing" not "writting".

Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>