Merge DT support for both 3.13 and 3.15 kernels
NB: This patch breaks backwards compatibility and requires
    kernel-3.13.0-0.rc7.34 or later.

In order to support 3.13 and 3.15 with the same firmware
version, the DT must provide device nodes expected by both
kernels. Both kernel use sata and menet nodes. Fortunately,
the the pre-upstream sata driver in 3.13 uses a different
compatibility string than the upstream sata driver in 3.15.
So we simply include both versions in the same DT. The
pre-upstream driver in 3.13 and the latest pre-upstream
driver for 3.15 use the same compatibilty string although
the actual contents of the nodes are different. We have to
rename the 3.13 menet node so the 3.15 driver will not try
to use it. We have a patched 3.13 kernel which looks for a
different compatibility string in addition to the string
previously used. We want to also support the old string in
3.13 so that the kernel is backwards compatible with earlier
firmware versions. However, the 3.13 kernel patch also looks
for a devid property which is only present in the old style
menet node. In this way, a 3.13 kernel will not use the menet
node meant for a 3.15 kernel.

Signed-off-by: Mark Salter <msalter@redhat.com>