[mustang] fix NOR flash partitioning
NOR flash is divided into a firmware area and two block devices.
Commit 1035be9d33edcbc "[mustang] enable building debug images"
increased image size so that it exceeded the firmware area and
encroached on both block devices. This patch reduces the size
of the last flash block device and increases the size of the
firmware area so that the image fits completely inside it.

Signed-off-by: Mark Salter <msalter@redhat.com>