reset enet clock in apm_enet_halt function
The apm_enet_halt function did not completely shut down the enet
hardware which allowed a spurious error message to sometimes be
delivered to kernel enet driver. This spurious error used a queue
id from the firmware and not known to the kernel. This caused the
kernel driver to get out of sync and eventually oops. This patch
does a clock and csr reset to ensure enet mac is completely
shut down by UEFI ExitBootServices handler.

Signed-off-by: Mark Salter <msalter@redhat.com>