fix setting of MAC address in ACPI table
The current code uses a ReadyToBootEvent to trigger code which fills
out the local-mac-address for ethenet device in the DSDT table. This
isn't triggering for some reason and APM's latest drop uses an
ExitBootServices event instead. The existing code also uses a "MAC0"
environment variable instead of "RGMII_MAC0". This patch pulls in
bits of APM's latest drop to use ExitBootServices and also adds
compatibility support if RGMII_MAC0 is used instead of MAC0.

Signed-off-by: Mark Salter <msalter@redhat.com>