Fix up ACPI tables for newer SATA and ENET kernel drivers
The existing ACPI tables support the older versions of the kernel
SATA and ENET drivers. Tweak them so that they work with the newer
upstream (or soon to be upstream in enet case) drivers without the
need for a lot of ugliness in the kernel.

Also fix a clock bug in ethernet ACPI object. The wrong bit was
being specified for menet device.

Signed-off-by: Mark Salter <msalter@redhat.com>