[mustang] remove rtcclk to prevent rtc-efi from breaking
The Mustang DTB describes an rtcclk that is otherwise unreferenced in the
DTB. It is apparently intended to be used by the XGene RTC driver, which
is not in the kernel. Since rtcclk isn't referenced, the Linux clock
framework powers its clock circuits down to save power. Unfortunately,
efi-rtc uses the same clock circuits, and shutting them down means that
reads and writes to efi-rtc fail.

This needs to be sorted out upstream. Remove rtcclk to make efi-rtc
functional while this is getting sorted out.

--Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>