Mustang: fix PCIe windows in DSDT
Currently the PCIe windows are marked ResourceConsumer which
means they are consumed by bridge itself. They should be marked
ResourceProducer because they are consumed by downstream devices.
That is, the windows are used to set downstream device BARs. This
has been working as-is because the kernel is not checking this,
but a future kernel patch will cause the kernel to only use the
ResourceProducer windows.

Signed-off-by: Mark Salter <msalter@redhat.com>