APMXGene-Mustang: lower high water mark for varstore GC by 7KB
Related bugzilla: 1151571

The ReclaimForOS() function in the variable driver is invoked at each
boot, in response to the Ready To Boot event. The function checks if the
remaining free space in the variable store can accommodate one maximally
sized new variable (give or take some metadata). If the room suffices,
then garbage collection is not done at that point.

Currently the APMXGene-Mustang platform description file inherits
PcdMaxVariableSize as 0x400 (1KB) from "MdeModulePkg/MdeModulePkg.dec".
This is problematic because it allows the OS to be booted with just one KB
of free space in the varstore, without attempting garbage collection. GC
is not done at runtime, which means that variable writes at runtime that
total over 1KB will run out of varstore and will be rejected, even though
the variable store might mostly be occupied by unreclaimed deleted
variables.

Raise PcdMaxVariableSize from 1KB to 8KB. Since the size of the variable
store, PcdFlashNvStorageVariableSize, is 0x20000 (128KB), this provides
for a high water mark of 120KB, rather than 127KB, and at least 8KB of
usable varstore space for the OS (unless the varstore is filled with live
variable data above that mark).

Signed-off-by: Laszlo Ersek <lersek@redhat.com>