APMXGeneArmPlatformLib: drop invalid memory allocation
"ArmPlatformPkg/APMXGenePkg/Library/APMXGeneArmPlatformLib" is a library
instance that is linked into
"ArmPlatformPkg/PrePeiCore/PrePeiCoreMPCore.inf", among other referers.
That client module has module type SEC, and in SEC memory allocation with
AllocateZeroPool() is unavailable. (This presents itself as a build
error.)

The root cause here is that the caller function,
ArmPlatformShowBoardBanner(), should definitely not be in a low-level
library that gets linked into a SEC module -- after all, this function is
only referenced from BDS
("ArmPlatformPkg/APMXGenePkg/Modules/ArmPlatformPkg/Bds/Bds.c").

In any case, for now we just replace the dynamic allocation with a block
on the stack (it's not big, FWSTRINGMAXLEN * 2 == 512).

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>