APMXGenePkg/AcpiPlatformDxe: fix PMU interrupt in Apic table
The template used to dynamically create the GIC struct in
the APIC table did not have the PerformaceInterruptGsiv field
set correctly (at all, really). This broke the perf events
driver in the kernel. Set the irq to match the Mustang hw
which uses a PPI (so we can just patch the template instead
of adding code to set it differently for each CPU).

Signed-off-by: Mark Salter <msalter@redhat.com>