mustang: fix PCI window in DSDT
The second QWordMemory definitionresource in the DSDT PCIe0 node is
incorrectly marked as "ResourceProducer" which means it is a window
from which downstream BARs may be assigned. In fact, that resource
is a reverse mapping from the PCI bus to RAM across the APM root
bridge. Since it can't be used to program downstream BARs and is
really used internally by the root bridge, it should be marked as
ResourceConsumer.

Signed-off-by: Mark Salter <msalter@redhat.com>